backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.7.0
chromadb==0.5.0
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
pymupdf==1.24.9
python-docx==1.1.2
pytesseract==0.3.10
pillow==10.4.0
requests==2.32.3
sentence-transformers==3.0.0
yagmail==0.15.293